Raise min deployment target to 7.0. Fix build problems.
FastMalloc.cpp: vm_page_shift is no longer a preprocessor macro, so it
cannot be used in some preprocessor checks done in this file before,
which has to be moved to a runtime check. The file contents is the file
contents from the WebKit at the tag Safari-600.1.4 (quite current),
which luckily compiles cleanly.

StructureInlines.h: Remove the inline for that function, which was
missing from the final library (seems that if an inline is always
inlined, the linker removes it from the object file?). Removing the
inline seems to fix it.

Raising the min deployment target avoids other linking problems related
to libc++ and libstdcpp. Seems that old deployment targets link with the
old library (since the new one isn’t available) which breaks the linking
in the final product (which only links with the new one).